chore: Refactor CopyButton and Tooltip to react-spa/shared
