+++ libs/portals/my-pages/information/src/components/PersonalInformation/Forms/ProfileForm/ProfileForm.tsx
@@ -6,19 +6,13 @@ import {
   GridColumn,
   GridContainer,
   GridRow,
-  Icon,
   Link,
   PhoneInput,
   SkeletonLoader,
   Text,
 } from '@island.is/island-ui/core'
 import { useLocale, useNamespaces } from '@island.is/localization'
-import {
-  LoadModal,
-  m,
-  parseNumber,
-  Tooltip,
-} from '@island.is/portals/my-pages/core'
+import { LoadModal, m, parseNumber } from '@island.is/portals/my-pages/core'
 import {
   useDeleteIslykillValue,
   useUserProfile,
@@ -27,7 +21,6 @@ import {
 import { useUserInfo } from '@island.is/react-spa/bff'
 import orderBy from 'lodash/orderBy'
 import { FormattedMessage } from 'react-intl'
-import copyToClipboard from 'copy-to-clipboard'
 import { useDelegationTypeFeatureFlag } from '../../../../hooks/useDelegationTypeFeatureFlag'
 import { useScopeAccess } from '../../../../hooks/useScopeAccess'
 import { emailsMsg, msg } from '../../../../lib/messages'
@@ -45,7 +38,7 @@ import { useConfirmNudgeMutation } from './confirmNudge.generated'
 import { DropModalType } from './types/form'
 import { InputEmail } from './components/Inputs/Email'
 import { AccessDenied } from '@island.is/portals/core'
-import { Problem } from '@island.is/react-spa/shared'
+import { CopyButton, Problem } from '@island.is/react-spa/shared'
 
 enum IdsUserProfileLinks {
   EMAIL = '/app/user-profile/email',
@@ -80,7 +73,6 @@ export const ProfileForm = ({
   const [internalLoading, setInternalLoading] = useState(false)
   const [showDropModal, setShowDropModal] = useState<DropModalType>()
   const [showEmailForm, setShowEmailForm] = useState(false)
-  const [copiedTraceSid, setCopiedTraceSid] = useState(false)
 
   const { deleteIslykillValue, loading: deleteLoading } =
     useDeleteIslykillValue()
@@ -191,13 +183,6 @@ export const ProfileForm = ({
 
   if (!isDelegationTypeEnabled && !hasUserProfileWriteScope) {
     return <AccessDenied />
-  }
-
-  const copy = (code?: string | null) => {
-    if (code) {
-      copyToClipboard(code)
-      setCopiedTraceSid(true)
-    }
   }
 
   return (
@@ -339,31 +324,7 @@ export const ProfileForm = ({
                 >
                   <Box display="flex">
                     <Text variant="eyebrow">{userInfo.profile.traceSid}</Text>
-                    <Tooltip
-                      text={
-                        copiedTraceSid
-                          ? formatMessage(m.copied)
-                          : formatMessage(m.copy)
-                      }
-                    >
-                      <Box
-                        cursor="pointer"
-                        marginLeft={1}
-                        display="flex"
-                        alignItems="center"
-                        justifyContent="center"
-                        onClick={() => {
-                          copy(userInfo.profile.traceSid)
-                        }}
-                      >
-                        <Icon
-                          icon="copy"
-                          type="outline"
-                          color="blue400"
-                          size="small"
-                        />
-                      </Box>
-                    </Tooltip>
+                    <CopyButton content={userInfo.profile.traceSid} />
                   </Box>
                 </InputSection>
               )}
